+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Experimental: Added Portable Dynamic Bone Collider component
 
 ### Fixed
+- [#626] Fixed an issue where an exception would be thrown if the base state corresponding to a synced layer override is destroyed.
 - [#607] Fixed an issue where zh-Hans l10n text fallback to zh-Hant.
 
 ### Changed